move to pglocks.go file
